+++ src/v3/handshake.rs
@@ -50,9 +50,6 @@ impl<Io> Handshake<Io> {
             io,
             shared,
             session: Some(st),
-            lw: 256,
-            read_hw: 4 * 1024,
-            write_hw: 4 * 1024,
             keepalive: Seconds(keepalive),
             return_code: mqtt::ConnectAckReason::ConnectionAccepted,
         }
@@ -65,9 +62,6 @@ impl<Io> Handshake<Io> {
             shared: self.shared,
             session: None,
             session_present: false,
-            lw: 256,
-            read_hw: 4 * 1024,
-            write_hw: 4 * 1024,
             keepalive: Seconds(30),
             return_code: mqtt::ConnectAckReason::IdentifierRejected,
         }
@@ -80,9 +74,6 @@ impl<Io> Handshake<Io> {
             shared: self.shared,
             session: None,
             session_present: false,
-            lw: 256,
-            read_hw: 4 * 1024,
-            write_hw: 4 * 1024,
             keepalive: Seconds(30),
             return_code: mqtt::ConnectAckReason::BadUserNameOrPassword,
         }
@@ -95,9 +86,6 @@ impl<Io> Handshake<Io> {
             shared: self.shared,
             session: None,
             session_present: false,
-            lw: 256,
-            read_hw: 4 * 1024,
-            write_hw: 4 * 1024,
             keepalive: Seconds(30),
             return_code: mqtt::ConnectAckReason::NotAuthorized,
         }
@@ -110,9 +98,6 @@ impl<Io> Handshake<Io> {
             shared: self.shared,
             session: None,
             session_present: false,
-            lw: 256,
-            read_hw: 4 * 1024,
-            write_hw: 4 * 1024,
             keepalive: Seconds(30),
             return_code: mqtt::ConnectAckReason::ServiceUnavailable,
         }
@@ -133,9 +118,6 @@ pub struct HandshakeAck<Io, St> {
     pub(crate) return_code: mqtt::ConnectAckReason,
     pub(crate) shared: Rc<MqttShared>,
     pub(crate) keepalive: Seconds,
-    pub(crate) lw: u16,
-    pub(crate) read_hw: u16,
-    pub(crate) write_hw: u16,
 }
 
 impl<Io, St> HandshakeAck<Io, St> {
@@ -147,51 +129,19 @@ impl<Io, St> HandshakeAck<Io, St> {
         self
     }
 
+    #[doc(hidden)]
+    #[deprecated(since = "0.7.6", note = "Use memory pool config")]
     #[inline]
     /// Set read/write buffer sizes
     ///
     /// By default max buffer size is 4kb for both read and write buffer,
     /// Min size is 256 bytes.
     pub fn buffer_params(
-        mut self,
-        max_read_buf: u16,
-        max_write_buf: u16,
-        min_buf_size: u16,
+        self,
+        _max_read_buf: u16,
+        _max_write_buf: u16,
+        _min_buf_size: u16,
     ) -> Self {
-        self.read_hw = max_read_buf;
-        self.write_hw = max_write_buf;
-        self.lw = min_buf_size;
-        self
-    }
-
-    #[doc(hidden)]
-    #[deprecated(since = "0.6.3")]
-    /// Set buffer low watermark size
-    ///
-    /// Low watermark is the same for read and write buffers.
-    /// By default lw value is 256 bytes.
-    pub fn low_watermark(mut self, lw: u16) -> Self {
-        self.lw = lw;
-        self
-    }
-
-    #[doc(hidden)]
-    #[deprecated(since = "0.6.3")]
-    /// Set read buffer high water mark size
-    ///
-    /// By default read hw is 4kb
-    pub fn read_high_watermark(mut self, hw: u16) -> Self {
-        self.read_hw = hw;
-        self
-    }
-
-    #[doc(hidden)]
-    #[deprecated(since = "0.6.3")]
-    /// Set write buffer high watermark size
-    ///
-    /// By default write hw is 4kb
-    pub fn write_high_watermark(mut self, hw: u16) -> Self {
-        self.write_hw = hw;
         self
     }
 }