+++ src/v3/shared.rs
@@ -2,7 +2,7 @@ use std::{cell::Cell, cell::RefCell, collections::VecDeque, num::NonZeroU16, rc:
 
 use ntex::channel::pool;
 use ntex::codec::{Decoder, Encoder};
-use ntex::util::{BytesMut, HashMap};
+use ntex::util::{BytesMut, HashMap, PoolId, PoolRef};
 
 use crate::error::{DecodeError, EncodeError};
 use crate::{io::State, types::packet_type, v3::codec};
@@ -23,11 +23,16 @@ pub(super) enum AckType {
 pub(super) struct MqttSinkPool {
     pub(super) queue: pool::Pool<Ack>,
     pub(super) waiters: pool::Pool<()>,
+    pub(super) pool: Cell<PoolRef>,
 }
 
 impl Default for MqttSinkPool {
     fn default() -> Self {
-        Self { queue: pool::new(), waiters: pool::new() }
+        Self {
+            queue: pool::new(),
+            waiters: pool::new(),
+            pool: Cell::new(PoolId::P5.pool_ref()),
+        }
     }
 }